Playing around with using nurseries and trio_asyncio
trio_asyncio seems to be necessary when using the newest prompt-toolkit
(especially to use prompt_async() but even when using just prompt()).

This still doesn't quite get the async stdout flow correct. I need to figure
out how to make the prompt not block stdout. I also will probably need to
figure out patch_stdout so it doesn't clobber the output, since we don't
actually know when a command is finished (it doesn't work as is, so we will
probably have to rewrite it to work async).